Update gimmeYourInfo.ps1
A lot of changes. The number of prompts that a user will receive will be less than or equal to one. A dictionary is no longer populated with MRU paths from recent entries. When I find a suitable test environment what will happen will be for the script to grab item one and move on.